Adds 404 error handling for API calls
Implements error handling for API calls to the Weaboo service.
Now, if an anime, manga, chapter, episode, or news item is not
found, the application will return a 404 error page, improving
the user experience by providing informative feedback when
content is unavailable.

Updates the manifest background and theme colors to black.