Implements wake lock to prevent screen dimming
Adds wake lock functionality to prevent the screen from dimming
or turning off while the application is in use. The wake lock is
automatically requested when the DOM is fully loaded and released
when the application is no longer in use.

Also, improves SEO by adding `anime` and `manga` keywords.